+++ lizard_fewsjdbc/migrations/0004_auto__add_field_jdbcsource_timezone_string.py
@@ -1,0 +1,49 @@
+# -*- coding: utf-8 -*-
+import datetime
+from south.db import db
+from south.v2 import SchemaMigration
+from django.db import models
+
+
+class Migration(SchemaMigration):
+
+    def forwards(self, orm):
+        # Adding field 'JdbcSource.timezone_string'
+        db.add_column('lizard_fewsjdbc_jdbcsource', 'timezone_string',
+                      self.gf('django.db.models.fields.CharField')(default='', max_length=40),
+                      keep_default=False)
+
+
+    def backwards(self, orm):
+        # Deleting field 'JdbcSource.timezone_string'
+        db.delete_column('lizard_fewsjdbc_jdbcsource', 'timezone_string')
+
+
+    models = {
+        'lizard_fewsjdbc.iconstyle': {
+            'Meta': {'object_name': 'IconStyle'},
+            'color': ('lizard_map.fields.ColorField', [], {'max_length': '8'}),
+            'fews_filter': ('django.db.models.fields.CharField', [], {'max_length': '40', 'null': 'True', 'blank': 'True'}),
+            'fews_location': ('django.db.models.fields.CharField', [], {'max_length': '40', 'null': 'True', 'blank': 'True'}),
+            'fews_parameter': ('django.db.models.fields.CharField', [], {'max_length': '40', 'null': 'True', 'blank': 'True'}),
+            'icon': ('django.db.models.fields.CharField', [], {'max_length': '40'}),
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'jdbc_source': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['lizard_fewsjdbc.JdbcSource']", 'null': 'True', 'blank': 'True'}),
+            'mask': ('django.db.models.fields.CharField', [], {'max_length': '40'})
+        },
+        'lizard_fewsjdbc.jdbcsource': {
+            'Meta': {'object_name': 'JdbcSource'},
+            'connector_string': ('django.db.models.fields.CharField', [], {'max_length': '200'}),
+            'customfilter': ('django.db.models.fields.TextField', [], {'null': 'True', 'blank': 'True'}),
+            'filter_tree_root': ('django.db.models.fields.CharField', [], {'max_length': '80', 'null': 'True', 'blank': 'True'}),
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'jdbc_tag_name': ('django.db.models.fields.CharField', [], {'max_length': '80'}),
+            'jdbc_url': ('django.db.models.fields.URLField', [], {'max_length': '200'}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '200'}),
+            'slug': ('django.db.models.fields.SlugField', [], {'max_length': '50'}),
+            'timezone_string': ('django.db.models.fields.CharField', [], {'default': "''", 'max_length': '40'}),
+            'usecustomfilter': ('django.db.models.fields.BooleanField', [], {'default': 'False'})
+        }
+    }
+
+    complete_apps = ['lizard_fewsjdbc']